Rename function to timeUntil
